Support for a short circuiting
Added support via the new option allowOverlayBars to disable the plugin when OS bars overlay the content. This corresponds with how most modern browsers handle scrollbars, it is often the effect you are going for when you use this plugin and if its already acceptable why run this code. Now you don't have to.